feat(ui): extract the JSON/guided-form pairing into editor-pair.js (#840)
View Definitions kept the two-way sync between CodeMirror and the guided
form inside vd-editor.js. That plumbing is now a shared host,
`window.HfsEditorPair.mount({textarea, view, grid, fields})`:

- form -> JSON as a single minimal, undoable CodeMirror transaction
  tagged with its origin, JSON -> form with a debounced re-render;
- the client-side "Invalid JSON" chip, whose message is captured once at
  mount so it survives server re-renders of the form card (the endpoint
  never sends `data-msg-invalid-json`);
- row <-> line cross-highlighting with reveal per pane;
- the textarea fallback when the CodeMirror bundle is missing;
- `host.fields`, extra fields editor-form.js sends with every request.

vd-editor.js keeps only what is FHIRPath-specific (injection, token
styles, server lint). code-editor.js gains the JSON token preset and the
`tabindex="0"` scroller for every mount; the shared-height CSS moves from
`#vd-editor-form` to the generic `.editor__doc`.

The `minimalChange` unit test moves to editor-pair.test.cjs and a new
editor-pair.spec.ts covers the chip after a guided-form round trip.
