+++ crates/ui/assets/code-editor.js
@@ -2,9 +2,9 @@
  * Shared CodeMirror 6 mount helper (#838, generalized out of
  * #753's original `vd-editor.js`).
  *
- * Every editor in this crate — the ViewDefinition JSON editor today, the SQL
- * pane editors — needs the exact same progressive-
- * enhancement contract over its `<textarea>`:
+ * Every editor in this crate — the ViewDefinition JSON editor, the SQL
+ * pane editors, the Library Details JSON editor (#840) — needs the exact
+ * same progressive-enhancement contract over its `<textarea>`:
  *
  *   - the textarea stays in the DOM and stays the form's source of truth,
  *     so Save/Duplicate/Enter (plain POSTs) keep submitting what the editor
@@ -14,15 +14,22 @@
  *     for any other script listening on the form;
  *   - the editor exposes the textarea's own `aria-label` on its own
  *     `role="textbox"` content, so the hidden textarea does not shadow it;
- *   - Tab is never captured for indentation — it keeps moving focus to the
- *     next form control, like it does over a plain textarea;
+ *   - the editor's own scroller is reachable by Tab even when it grows tall
+ *     enough to scroll internally (`contentAttributes tabindex="0"`, #840 —
+ *     axe's `scrollable-region-focusable` check does not credit a bare
+ *     `contenteditable` as focusable content for its scrolling ancestor);
+ *   - Tab itself is never captured for indentation — it keeps moving focus
+ *     to the next form control, like it does over a plain textarea;
  *   - the wrapper and the `EditorView` are built fully in memory and the
  *     live DOM is touched only once both succeed, so a construction error
  *     never leaves a hidden textarea with no editor to show for it.
  *
  * What differs per editor — the language grammar, syntax highlighting,
  * lint, fold — is not this helper's concern: callers pass it all in via
- * `options`. This file only owns the wrapper/sync/back-out plumbing above.
+ * `options`, except the JSON token-color palette every JSON-editing page
+ * shares (`jsonHighlight()` below, #840) — one preset rather than a copy
+ * per caller. This file only owns the wrapper/sync/back-out plumbing above
+ * and that shared preset.
  *
  * `window.HfsCodeEditor.mount(textarea, options)` returns the `EditorView`
  * it built, or `null` if `window.HfsCodeMirror` (the vendored bundle)
@@ -41,6 +48,13 @@
  *   - `wrapperClass` — extra class name(s) on the wrapper, alongside the
  *                   shared `code-editor` class every mount gets.
  *   - `id`        — id attribute for the wrapper element.
+ *
+ * `window.HfsCodeEditor.jsonHighlight()` returns a `HighlightStyle`,
+ * scoped to `window.HfsCodeMirror.jsonLanguage`, coloring the outer JSON's
+ * five token classes (`cmt-json-key`/`-string`/`-number`/`-literal`/
+ * `-punct`, `app.css`) — pass it through `syntaxHighlighting(...)` in
+ * `options.highlight` the same as any other `HighlightStyle`. `null` if
+ * `window.HfsCodeMirror` is not loaded.
  */
 (function () {
   "use strict";
@@ -87,6 +101,16 @@
         CM.EditorView.contentAttributes.of({
           "aria-label": textarea.getAttribute("aria-label") || "",
         }),
+        // `.cm-content` is `contenteditable`, genuinely reachable by Tab in
+        // every real browser without this — but axe's own
+        // `scrollable-region-focusable` check does not credit a bare
+        // `contenteditable` as "focusable content" for its *ancestor*
+        // `.cm-scroller` (`overflow: auto` in the shared chrome, #838) —
+        // only an element that itself carries an explicit `tabindex`. A
+        // separate `contentAttributes.of` call merges its key into the one
+        // above via CodeMirror's own facet combination (#840, generalized
+        // off #838's ViewDefinition-only copy of this rule).
+        CM.EditorView.contentAttributes.of({ tabindex: "0" }),
         // No indentWithTab: Tab must keep moving focus to the next form
         // control, not indent inside the editor.
         CM.keymap.of(
@@ -124,5 +148,27 @@
     }
   }
 
-  window.HfsCodeEditor = { mount: mount };
+  /* The JSON token-color preset every JSON-editing page shares (#840,
+   * lifted verbatim out of the ViewDefinition editor's own copy): classes
+   * only, every actual color lives in `app.css` as a CSS variable, scoped
+   * to `jsonLanguage` so it only ever paints the outer JSON grammar, never
+   * a language injected into one of its string values (`vd-editor.js`'s own
+   * FHIRPath HighlightStyle stays separate and scoped to its own
+   * language for exactly that reason). */
+  function jsonHighlight() {
+    var CM = window.HfsCodeMirror;
+    if (!CM) return null;
+    return CM.HighlightStyle.define(
+      [
+        { tag: CM.tags.propertyName, class: "cmt-json-key" },
+        { tag: CM.tags.string, class: "cmt-json-string" },
+        { tag: CM.tags.number, class: "cmt-json-number" },
+        { tag: [CM.tags.bool, CM.tags.null], class: "cmt-json-literal" },
+        { tag: [CM.tags.separator, CM.tags.squareBracket, CM.tags.brace], class: "cmt-json-punct" },
+      ],
+      { scope: CM.jsonLanguage }
+    );
+  }
+
+  window.HfsCodeEditor = { mount: mount, jsonHighlight: jsonHighlight };
 })();
