+++ crates/ui/e2e/unit/editor-pair.test.cjs
@@ -1,20 +1,21 @@
 const test = require("node:test");
 const assert = require("node:assert/strict");
 
-const vdEditor = require("../../assets/vd-editor.js");
+const editorPair = require("../../assets/editor-pair.js");
 
-// #843: the guided form pushes its edits into CodeMirror as one minimal
-// transaction (common-prefix/common-suffix diff), so the caret, the scroll
-// position, and the undo history all behave like a targeted manual edit
-// instead of a full-document replace.
+// #843 (extracted onto editor-pair.js in #840): the guided form pushes its
+// edits into CodeMirror as one minimal transaction (common-prefix/
+// common-suffix diff), so the caret, the scroll position, and the undo
+// history all behave like a targeted manual edit instead of a
+// full-document replace.
 
 test("identical text produces no change (no transaction)", () => {
-  assert.equal(vdEditor.minimalChange("same", "same"), null);
-  assert.equal(vdEditor.minimalChange("", ""), null);
+  assert.equal(editorPair.minimalChange("same", "same"), null);
+  assert.equal(editorPair.minimalChange("", ""), null);
 });
 
 test("an insertion at the end is a single trailing insert", () => {
-  assert.deepEqual(vdEditor.minimalChange("abc", "abcdef"), {
+  assert.deepEqual(editorPair.minimalChange("abc", "abcdef"), {
     from: 3,
     to: 3,
     insert: "def",
@@ -22,7 +23,7 @@ test("an insertion at the end is a single trailing insert", () => {
 });
 
 test("an insertion at the start is a single leading insert", () => {
-  assert.deepEqual(vdEditor.minimalChange("def", "abcdef"), {
+  assert.deepEqual(editorPair.minimalChange("def", "abcdef"), {
     from: 0,
     to: 0,
     insert: "abc",
@@ -30,7 +31,7 @@ test("an insertion at the start is a single leading insert", () => {
 });
 
 test("a replacement in the middle keeps the untouched prefix and suffix out of the range", () => {
-  assert.deepEqual(vdEditor.minimalChange("abcXdef", "abcYYdef"), {
+  assert.deepEqual(editorPair.minimalChange("abcXdef", "abcYYdef"), {
     from: 3,
     to: 4,
     insert: "YY",
@@ -38,7 +39,7 @@ test("a replacement in the middle keeps the untouched prefix and suffix out of t
 });
 
 test("a deletion is an empty insert over the removed range", () => {
-  assert.deepEqual(vdEditor.minimalChange("abcXYZdef", "abcdef"), {
+  assert.deepEqual(editorPair.minimalChange("abcXYZdef", "abcdef"), {
     from: 3,
     to: 6,
     insert: "",
@@ -46,7 +47,7 @@ test("a deletion is an empty insert over the removed range", () => {
 });
 
 test("replacing the whole document (no shared prefix or suffix) covers the full range", () => {
-  assert.deepEqual(vdEditor.minimalChange("abc", "xyz"), {
+  assert.deepEqual(editorPair.minimalChange("abc", "xyz"), {
     from: 0,
     to: 3,
     insert: "xyz",
@@ -57,7 +58,7 @@ test("a shared prefix and suffix that overlap in the shorter text do not double-
   // "aaa" -> "aa": every character in "aa" matches as both a prefix and (if
   // unguarded) a suffix of "aaa" - the shared region must not be counted
   // twice, which would put `to` before `from`.
-  const change = vdEditor.minimalChange("aaa", "aa");
+  const change = editorPair.minimalChange("aaa", "aa");
   assert.ok(change.from <= change.to, `from (${change.from}) must not exceed to (${change.to})`);
   // Re-applying the change must reproduce the target text exactly.
   const applied = "aaa".slice(0, change.from) + change.insert + "aaa".slice(change.to);
@@ -67,7 +68,7 @@ test("a shared prefix and suffix that overlap in the shorter text do not double-
 test("pretty-printing a document (whitespace-only change) is still a minimal range, not a full replace", () => {
   const before = '{"a":1}';
   const after = '{\n  "a": 1\n}';
-  const change = vdEditor.minimalChange(before, after);
+  const change = editorPair.minimalChange(before, after);
   // The literal "1" in the middle is untouched text on both sides, so the
   // diff must not span the entire document even though nearly every
   // character differs once whitespace is inserted around it.
